Get-IcingaCheckCommandConfig: improve -IcingaConfig documentation
The old Cmdlet's documentation implied that just by passing the
-IcingaConfig switch, an Icinga 2 configuration would be generated.
However, the first two examples does not work as intended, as
-IcingaConfig only works together with the -OutDirectory parameter.

The documentation was updated to reflect as Get-IcingaCheckCommandConfig
behaves and how one can generate Icinga 2 configurations. In addition,
a note was added to the Cmdlet's API documentation to handle
expectations of the switch.