Set maximum cache duration for service daemons to the right value
This commit sets the maximum duration for cached data for service
daemons to the proper amount.

Previously it was multiplied by 60, probably a typo from a time
when the assigned value was in minutes and not in seconds directly.

Therefore the value was 60 times to high which leads to a massive amount
of cached data which drags down the performance significantly.